requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
coreapi==2.3.3
coreschema==0.0.4
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
PyJWT==2.10.1
pytz==2024.2
requests==2.32.3
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
sqlparse==0.5.3
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
